+++ src/main/java/ai/iambot/elasticsearch/script/metrics/CosineMetricProvider.java
@@ -2,7 +2,7 @@ package ai.iambot.elasticsearch.script.metrics;
 
 public class CosineMetricProvider implements MetricProvider{
     @Override
-    public Metric getMetric(double[] inputVector) {
+    public Metric getMetric(float[] inputVector) {
         return new CosineMetric(inputVector);
     }
 }
