+++ src/main/java/ai/iambot/elasticsearch/script/metrics/MetricProvider.java
@@ -2,5 +2,5 @@ package ai.iambot.elasticsearch.script.metrics;
 
 public interface MetricProvider {
 
-    Metric getMetric(double[] inputVector);
+    Metric getMetric(float[] inputVector);
 }
