+++ src/main/java/ai/iambot/elasticsearch/script/metrics/TSSSMetric.java
@@ -1,13 +1,13 @@
 package ai.iambot.elasticsearch.script.metrics;
 
 public class TSSSMetric implements Metric{
-    double[] inputVector;
-    public TSSSMetric(double[] inputVector){
+    float[] inputVector;
+    public TSSSMetric(float[] inputVector){
         this.inputVector = inputVector;
     }
 
     @Override
-    public double metric(double[] queryVector) {
+    public float metric(float[] queryVector) {
         return TSSSHelper.TS_SS(this.inputVector, queryVector);
     }
 }
